components/frontend-api/requirements.txt
google-cloud-firestore
cachetools==4.2.0
